requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.1
pymongo==3.7.2
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
Werkzeug==0.14.1
